Create excpetions class for invalid user issues
Fixes#71
