+++ src/main/java/seedu/duke/InvalidAgeException.java
@@ -5,7 +5,7 @@ package seedu.duke;
  * This exception is thrown if the add housekeeper command has a given age which is not an integer.
  */
 public class InvalidAgeException extends HotelLiteManagerException {
-    private static final String ERROR_MESSAGE = "Error! AGE must be integer";
+    private static final String ERROR_MESSAGE = "Error! AGE must be integer.";
 
     @Override
     public String getErrorMessage() {
